+++ cpp/dolfin/io/XDMFFile.cpp
@@ -37,7 +37,7 @@
 #include <dolfin/mesh/MeshValueCollection.h>
 #include <dolfin/mesh/Partitioning.h>
 #include <dolfin/mesh/Vertex.h>
-#include <dolfin/mesh/utils.h>
+#include <dolfin/mesh/cell_types.h>
 #include <iomanip>
 #include <memory>
 #include <petscvec.h>
