+++ cpp/dolfin/mesh/dolfin_mesh.h
@@ -26,4 +26,5 @@ namespace dolfin
 #include <dolfin/mesh/Partitioning.h>
 #include <dolfin/mesh/Topology.h>
 #include <dolfin/mesh/Vertex.h>
+#include <dolfin/mesh/cell_types.h>
 #include <dolfin/mesh/utils.h>
